Use semver when comparing latest and current versions
If compiling 1.3.0 from source while the latest release is 1.2.0,
according to https://noaa-apt.mbernardi.com.ar/version_check
the user would get a notification/warning that an update is
available, even if the update is an "older" version:
"Version 1.2.0 is available for download!".

This fix parses the current / latest using semantic versioning and
does not suggest an update if current >= latest.